requirements.txt
pycryptodome>=3.15.0
pyinstaller>=5.0.0

# built-in: base64, zlib, marshal, ast, sys, os, re
# pyinstaller: for creating standalone executables
